moved database settings and fill_database function to db_config file
