+++ shop_bot/shop_main.py
@@ -4,7 +4,7 @@ import sys
 
 from aiogram import Bot, Dispatcher
 
-from shop_bot.database.models import async_main
+from shop_bot.database.db_config import async_main
 from shop_bot.handlers import router
